content/canvas/test/test_size.attributes.type.get.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.type.get</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.width === 120, "canvas.width === 120");
ok(canvas.height === 60, "canvas.height === 60");
ok(canvas.getAttribute('width') === '120', "canvas.getAttribute('width') === '120'");
ok(canvas.getAttribute('height') === '60', "canvas.getAttribute('height') === '60'");

if (!_deferred) SimpleTest.finish();
});
</script>
